Remove 1 step latency between state and prediction
CANs take 1 step to propagate information for one step. This can introduce delay between when a state is seen, and when it is used in the prediction. By rearranging the order in which information is processed, now CANs can use the information in one step linearly. Using the information using non-linear features still requires multiple time-steps (and is by design)